change return type from int to long for result of int multiplication

+++ Graffiti_Core/org/Vector2i.java
@@ -122,7 +122,7 @@ public class Vector2i {
 	/**
 	 * @return x*y
 	 */
-	public int getArea() {
+	public long getArea() {
 		return x * y;
 	}
 }
